Move exit diagnostics to attempt level
